pyproject.toml
[tool.poetry]
name = "todo-api"
version = "0.1.0"
description = "Todo API to manage simple tasks"
authors = [ "Diogo Faria <diogofaria073@gmail.com>" ]
readme = "README.md"


[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.115.8"
ruff = "^0.1.0"
sqlalchemy = "^2.0.38"
pydantic-settings = "^2.8.0"
alembic = "^1.14.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"


[tool.pytest.ini_options]
python_path = "."
addopts = "-p no:warnings"


[tool.ruff.lint]
preview = true
select = [ "I", "F", "E", "W", "PL", "PT" ]


[tool.ruff.format]
preview = true
quote-style = "single"


[tool.taskipy.tasks]
run = "fastapi dev todo_api/app.py"
pre_test = "task lint"
test = "pytest --cov=todo_api -vv"
post_test = "coverage html"
lint = "ruff check . && ruff check . --diff"
format = "ruff check . --fix && ruff format ."
create_migrations = "alembic revision --autogenerate -m 'create migration'"
apply_migrations = "alembic upgrade head"


[toll.ruff]
line-length = 79
extend-exclude = [ "todo_api/repositories/migrations", "./poetry.lock" ]


[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
